requirements.txt
google-genai
langchain-google-genai
faiss-cpu
PyPDF2
numpy
python-dotenv
